[VER] : 1차 배포준비 (QA테스트용 서버버전)

+++ src/main/java/com/finalproject/backend/baseballmate/join/JoinRequestsService.java
@@ -48,7 +48,7 @@ public class JoinRequestsService {
             {
                 joinRequestsRepository.save(joinRequests);
                 AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
-                String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님이 내가 만든 모임 : " +group.getTitle()+" 에 참여신청을 하셨습니다.";
+                String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 모임 : " +group.getTitle()+" 에 *참여신청을 하셨습니다.";
                 alarmRequestDto.setUserId(user.getId());
                 alarmRequestDto.setContents(signupAlarm);
                 alarmRequestDto.setJoinRequestId(joinRequests.getId());
@@ -63,7 +63,7 @@ public class JoinRequestsService {
                 if (allChatInfoQueryRepository.findByChatRoom_IdAndUser_Id(roomId, userId) == null) {
                     joinRequestsRepository.save(joinRequests);
                     AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
-                    String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님이 내가 만든 모임 : " +group.getTitle()+" 에 참여신청을 하셨습니다.";
+                    String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 모임 : " +group.getTitle()+" 에 *참여신청을 하셨습니다.";
                     alarmRequestDto.setUserId(user.getId());
                     alarmRequestDto.setContents(signupAlarm);
                     alarmRequestDto.setJoinRequestId(joinRequests.getId());
@@ -132,7 +132,6 @@ public class JoinRequestsService {
         Long postId = joinRequests.getPostId();
         if (!tOrF) {
             joinRequestsRepository.delete(joinRequests);
-            groupService.cancelApplication2(postId,joinRequests);
             return "거절되었습니다";
         }
         User user = userRepository.findById(joinRequests.getUserId()).orElseThrow(
@@ -147,7 +146,7 @@ public class JoinRequestsService {
         groupService.applyGroup2(postId,joinRequests);
         return "승인되었습니다";
 
-        }
+    }
 
 
     // 채팅방 인원수 제한
@@ -166,10 +165,31 @@ public class JoinRequestsService {
     // 채팅방 입장 신청 취소
     @Transactional
     public String requestJoinCancel(UserDetailsImpl userDetails, Long joinId) {
-        Long userId = userDetails.getUser().getId();
-        JoinRequests joinRequests = joinRequestsRepository.findByUserIdAndPostId(userId, joinId);
-        joinRequestsRepository.delete(joinRequests);
-        return "참가신청 취소가 완료되었습니다";
+        User user = userDetails.getUser();
+
+        JoinRequests joinRequests = joinRequestsRepository.findByUserIdAndPostId(user.getId(), joinId);
+        if(joinRequests==null)
+        {
+            Group group =groupRepository.findByGroupId(joinId);
+            groupService.cancelApplication(group.getGroupId(),userDetails);
+            return "참가했던 모임에서의 취소가 완료되었습니다";
+        }
+
+        else{
+            Group group=groupRepository.findByGroupId(joinRequests.getId());
+            joinRequestsRepository.delete(joinRequests);
+            AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
+            String signupAlarm = group.getCreatedUser().getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 모임 : " +group.getTitle()+" 에 했던 *신청을 취소 하셨습니다.";
+            alarmRequestDto.setUserId(group.getCreatedUser().getId());
+            alarmRequestDto.setContents(signupAlarm);
+            alarmRequestDto.setJoinRequestId(joinRequests.getId());
+            alarmRequestDto.setAlarmType("Normal");
+            alarmService.createAlarm(alarmRequestDto);
+
+            return "참가신청 취소가 완료되었습니다";
+        }
+
+
     }
 
     // 게시글 삭제시 승인대기 목록 삭제
@@ -196,7 +216,7 @@ public class JoinRequestsService {
             {
                 joinRequestsRepository.save(joinRequests);
                 AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
-                String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님이 내가 만든 스크린 야구모임 : " +group.getTitle()+" 에 참여신청을 하셨습니다.";
+                String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 스크린 야구모임 : " +group.getTitle()+" 에 *참여신청을 하셨습니다.";
                 alarmRequestDto.setUserId(user.getId());
                 alarmRequestDto.setContents(signupAlarm);
                 alarmRequestDto.setJoinRequestId(joinRequests.getId());
@@ -211,7 +231,7 @@ public class JoinRequestsService {
                 if (allChatInfoQueryRepository.findByChatRoom_IdAndUser_Id(roomId, userId) == null) {
                     joinRequestsRepository.save(joinRequests);
                     AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
-                    String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님이 내가 만든 스크린 야구모임 : " +group.getTitle()+" 에 참여신청을 하셨습니다.";
+                    String signupAlarm = user.getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 스크린 야구모임 : " +group.getTitle()+" 에 *참여신청을 하셨습니다.";
                     alarmRequestDto.setUserId(user.getId());
                     alarmRequestDto.setContents(signupAlarm);
                     alarmRequestDto.setJoinRequestId(joinRequests.getId());
@@ -280,7 +300,6 @@ public class JoinRequestsService {
         Long postId = joinRequests.getPostId();
         if (!tOrF) {
             joinRequestsRepository.delete(joinRequests);
-            screenService.cancleApplication2(postId,joinRequests);
             return "거절되었습니다";
         }
         User user = userRepository.findById(joinRequests.getUserId()).orElseThrow(
@@ -301,11 +320,31 @@ public class JoinRequestsService {
     // 채팅방 입장 신청 취소
     @Transactional
     public String requestJoinCancel2(UserDetailsImpl userDetails, Long joinId) {
-        Long userId = userDetails.getUser().getId();
-        JoinRequests joinRequests = joinRequestsRepository.findByUserIdAndPostId(userId, joinId);
-        joinRequestsRepository.delete(joinRequests);
-        return "참가신청 취소가 완료되었습니다";
-    }
+        User user = userDetails.getUser();
 
+        JoinRequests joinRequests = joinRequestsRepository.findByUserIdAndPostId(user.getId(), joinId);
+        if(joinRequests==null)
+        {
+            Screen group =screenRepository.findByScreenId(joinId);
+            screenService.cancleApplication(group.getScreenId(),userDetails);
+            return "참가했던 모임에서의 취소가 완료되었습니다";
+        }
+
+        else{
+            Screen group=screenRepository.findByScreenId(joinRequests.getId());
+            joinRequestsRepository.delete(joinRequests);
+            AlarmRequestDto alarmRequestDto = new AlarmRequestDto();
+            String signupAlarm = group.getScreenCreatedUser().getUsername() + "님! "+userDetails.getUser().getUsername()+" 님께서* 내가 만든 모임 : " +group.getTitle()+" 에 했던 *신청을 취소 하셨습니다.";
+            alarmRequestDto.setUserId(group.getScreenCreatedUser().getId());
+            alarmRequestDto.setContents(signupAlarm);
+            alarmRequestDto.setJoinRequestId(joinRequests.getId());
+            alarmRequestDto.setAlarmType("Normal");
+            alarmService.createAlarm(alarmRequestDto);
+
+            return "참가신청 취소가 완료되었습니다";
+        }
+
+
+    }
 
 }
